+++ src/cibmangotree/gui/components/__init__.py
@@ -1,15 +1,10 @@
 from .analysis import AnalysisParamsCard
 from .analysis_utils import analysis_label, present_timestamp
 from .choice_fork import two_button_choice_fork_content
+from .exit_confirmation import ExitConfirmationDialog
 from .import_options import ImportOptionsDialog
 from .manage_analyses import ManageAnalysisDialog
 from .manage_projects import ManageProjectsDialog
-from .stepper_steps import (
-    AnalyzerSelectionStep,
-    ColumnMappingStep,
-    ParamsConfigStep,
-    RunAnalysisStep,
-)
 from .toggle import ToggleButton, ToggleButtonGroup
 from .upload.upload_button import UploadButton
 
@@ -21,10 +16,7 @@ __all__ = [
     "ManageAnalysisDialog",
     "ManageProjectsDialog",
     "ImportOptionsDialog",
-    "AnalyzerSelectionStep",
-    "ColumnMappingStep",
-    "ParamsConfigStep",
-    "RunAnalysisStep",
+    "ExitConfirmationDialog",
     "analysis_label",
     "present_timestamp",
     "two_button_choice_fork_content",
